requirements.txt
streamlit
google-generativeai
pandas
Pillow
python-dotenv
python-docx
markdown
